Flatten the tos page
