+++ src/App.tsx
@@ -3,8 +3,8 @@ import React, { Suspense } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 const AppLayout = React.lazy(() => import("./components/layout/AppLayout"));
-const Tos = React.lazy(() => import("./components/upload/Tos"));
-const New = React.lazy(() => import("./components/upload/Tos/New"));
+const Tos = React.lazy(() => import("./components/pages/tos"));
+const New = React.lazy(() => import("./components/pages/tos/New"));
 const Result = React.lazy(() => import("./components/tree/Result"));
 const NotFound = React.lazy(() => import("./components/NotFound"));
 const History = React.lazy(() => import("./components/pages/history"));
